Using array destructuring.
This seems to be the cleanest way to deal with the fact that typescript
won't do the spread operator unless you allow arbitrary number of
parameters in the function definition.

Why they don't allow this, I'm not entirely sure. At first, I thought it
was because they couldn't guarantee that the input array was of the
correct length. This is not true. The generated code for destructuring
doesn't change if the inputs are wrong. So I'm at a bit of a loss why they
support array destructuring, but not the spread operator.